FFI: Move remaining store-based APIs over to bridge_fn
The exception is signal_sealed_session_cipher_decrypt, which like
signal_identitykeypair_deserialize has multiple return values.
